Homework 10 is done

+++ src/main/java/ru/utmn/dyagunov/tax_subsidies/model/TaxSubsidy.java
@@ -30,7 +30,7 @@ public class TaxSubsidy {
     @Column()
     Integer timePeriod;
     @Column()
-    Float observationValue;
+    Double observationValue;
     @Column()
     String regimeName;
